backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
requests==2.31.0
httpx==0.25.1
pytest==7.4.3
fastapi[test]